change price type in product and price history entities

+++ ProductService.Domain/PriceHistory.cs
@@ -6,8 +6,8 @@ public class PriceHistory
 {
     public Guid Id { get; set; }
     public Guid ProductId { get; set; }
-    public double OldPrice { get; set; }
-    public double NewPrice { get; set; }
+    public decimal OldPrice { get; set; }
+    public decimal NewPrice { get; set; }
     public ActionType Action { get; set; }
     public DateTimeOffset Timestamp { get; set; }
     public Product Product { get; set; }
